Dummy consultancy
The files contain a detailed explanation of our acting as a dummy consultancy named "Bold Consultancy". In this project, it's aimed to provide a well-researched analysis for a US steel manufacturer "Dorman Products", who seeks our services to expand their operations in other parts of world and we finally conclude with M&A considering all factors and competitive steel market.